[shaderpack] Add nova shaderpack renderpass loading
